Fix travis tests as it doesn't have s3 access
